# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_heartgryphon_dsp_internal_audit.Entry:
    properties:
      action:
        type: string
      actor:
        type: string
      advertiser_id:
        type: integer
      created_at:
        type: string
      details:
        additionalProperties: {}
        type: object
      id:
        type: integer
      resource_id:
        type: integer
      resource_type:
        type: string
    type: object
  github_com_heartgryphon_dsp_internal_billing.Transaction:
    properties:
      advertiser_id:
        type: integer
      amount_cents:
        type: integer
      balance_after:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      reference_id:
        type: string
      type:
        type: string
    type: object
...
host: localhost:8181
info:
  contact: {}
  description: Demand-Side Platform — programmatic advertising API
  title: DSP Platform API
  version: "1.0"
paths:
...
  /analytics/snapshot:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get analytics snapshot
      tags:
      - analytics
  /analytics/stream:
    get:
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Real-time analytics SSE stream
      tags:
      - analytics
  /audit-log:
    get:
      parameters:
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_heartgryphon_dsp_internal_audit.Entry'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get my audit log
      tags:
      - audit
...
  /billing/balance/{id}:
    get:
      parameters:
      - description: Advertiser ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              advertiser_id:
                type: integer
              balance_cents:
                type: integer
              billing_type:
                type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get advertiser balance
      tags:
      - billing
  /billing/topup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Top-up data
        in: body
        name: body
        required: true
        schema:
          properties:
            advertiser_id:
              type: integer
            amount_cents:
              type: integer
            description:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_heartgryphon_dsp_internal_billing.Transaction'
      security:
      - ApiKeyAuth: []
      summary: Top up advertiser balance
      tags:
      - billing
  /billing/transactions:
    get:
      parameters:
      - description: Advertiser ID
        in: query
        name: advertiser_id
        type: integer
      - default: 50
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_heartgryphon_dsp_internal_billing.Transaction'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List billing transactions
      tags:
      - billing
...
  /export/campaign/{id}/bids:
    get:
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - ApiKeyAuth: []
      summary: Export bid details as CSV
      tags:
      - export
  /export/campaign/{id}/stats:
    get:
      parameters:
      - description: Campaign ID
        in: path
        name: id
        required: true
        type: integer
      - description: Start date
        in: query
        name: from
        type: string
      - description: End date
        in: query
        name: to
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - ApiKeyAuth: []
      summary: Export campaign stats as CSV
      tags:
      - export
...
securityDefinitions:
  AdminAuth:
    in: header
    name: X-Admin-Token
    type: apiKey
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"